Refactor HTTP routes
