+++ Sources/Server/Controllers/ScheduleController.swift
@@ -1,0 +1,23 @@
+//
+//  ScheduleController.swift
+//  Shuttle Tracker Server
+//
+//  Created by Gabriel Jacoby-Cooper on 1/10/24.
+//
+
+import Vapor
+
+/// A structure that registers routes for the schedule.
+/// - Important: This structure registers a route on the index path of its provided routes builder, so make sure to enclose it in a named routes group to avoid path collisions.
+/// - Remark: In the context of this structure, the term “route” refers to an HTTP route, not a shuttle route.
+struct ScheduleController: RouteCollection {
+	
+	func boot(routes: any RoutesBuilder) throws {
+		routes.get(use: self.index(_:))
+	}
+	
+	private func index(_ request: Request) -> Response {
+		return request.redirect(to: "/schedule.json")
+	}
+	
+}
